+++ tests/testthat/test-cv_nested_surv_glmnet_cox.R
@@ -12,11 +12,21 @@ param_list_glmnet <- expand.grid(
   alpha = seq(0, 1, .2)
 )
 glmnet_bounds <- list(alpha = c(0., 1.))
+
+
+ncores <- ifelse(
+  test = parallel::detectCores() > 4,
+  yes = 4L,
+  no = ifelse(
+    test = parallel::detectCores() < 2L,
+    yes = 1L,
+    no = parallel::detectCores()
+  )
+)
 optim_args <- list(
-  iters.n = 2L,
+  iters.n = ncores,
   kappa = 3.5,
-  acq = "ucb",
-  otherHalting = list(timeLimit = 30)
+  acq = "ucb"
 )
 
 split_vector <- splitTools::multi_strata(
@@ -50,31 +60,31 @@ test_that(
   desc = "test nested cv, bayesian - surv_glmnet_cox",
   code = {
 
-    surv_glmnet_cox_cv <- MLNestedCV$new(
+    surv_glmnet_cox_optimization <- MLNestedCV$new(
       learner = learner,
       strategy = "bayesian",
       fold_list = fold_list,
       k_tuning = 3L,
-      ncores = 2L,
+      ncores = ncores,
       seed = seed
     )
 
-    surv_glmnet_cox_cv$parameter_bounds <- glmnet_bounds
-    surv_glmnet_cox_cv$parameter_grid <- param_list_glmnet
-    surv_glmnet_cox_cv$split_type <- "stratified"
-    surv_glmnet_cox_cv$split_vector <- split_vector
-    surv_glmnet_cox_cv$optim_args <- optim_args
+    surv_glmnet_cox_optimization$parameter_bounds <- glmnet_bounds
+    surv_glmnet_cox_optimization$parameter_grid <- param_list_glmnet
+    surv_glmnet_cox_optimization$split_type <- "stratified"
+    surv_glmnet_cox_optimization$split_vector <- split_vector
+    surv_glmnet_cox_optimization$optim_args <- optim_args
 
     # set data
-    surv_glmnet_cox_cv$set_data(
+    surv_glmnet_cox_optimization$set_data(
       x = train_x,
       y = train_y
     )
 
-    cv_results <- surv_glmnet_cox_cv$execute()
+    cv_results <- surv_glmnet_cox_optimization$execute()
     expect_type(cv_results, "list")
     expect_equal(dim(cv_results), c(5, 3))
-    expect_true(inherits(x = surv_glmnet_cox_cv$results, what = "mlexCV"))
+    expect_true(inherits(x = surv_glmnet_cox_optimization$results, what = "mlexCV"))
   }
 )
 
@@ -83,30 +93,30 @@ test_that(
   desc = "test nested cv, grid - surv_glmnet_cox",
   code = {
 
-    surv_glmnet_cox_cv <- MLNestedCV$new(
+    surv_glmnet_cox_optimization <- MLNestedCV$new(
       learner = learner,
       strategy = "grid",
       fold_list = fold_list,
       k_tuning = 3L,
-      ncores = 2L,
+      ncores = ncores,
       seed = seed
     )
 
-    surv_glmnet_cox_cv$parameter_grid <- param_list_glmnet
-    surv_glmnet_cox_cv$split_type <- "stratified"
-    surv_glmnet_cox_cv$split_vector <- split_vector
-    surv_glmnet_cox_cv$optim_args <- optim_args
+    surv_glmnet_cox_optimization$parameter_grid <- param_list_glmnet
+    surv_glmnet_cox_optimization$split_type <- "stratified"
+    surv_glmnet_cox_optimization$split_vector <- split_vector
+    surv_glmnet_cox_optimization$optim_args <- optim_args
 
     # set data
-    surv_glmnet_cox_cv$set_data(
+    surv_glmnet_cox_optimization$set_data(
       x = train_x,
       y = train_y
     )
 
-    cv_results <- surv_glmnet_cox_cv$execute()
+    cv_results <- surv_glmnet_cox_optimization$execute()
     expect_type(cv_results, "list")
     expect_equal(dim(cv_results), c(5, 3))
-    expect_true(inherits(x = surv_glmnet_cox_cv$results, what = "mlexCV"))
+    expect_true(inherits(x = surv_glmnet_cox_optimization$results, what = "mlexCV"))
   }
 )