+++ tests/testthat/test-cv_nested_surv_xgboost_cox.R
@@ -14,18 +14,28 @@ param_list_xgboost <- expand.grid(
   subsample = seq(0.6, 1, .2),
   colsample_bytree = seq(0.6, 1, .2),
   min_child_weight = seq(1, 5, 4),
-  learning_rate = seq(0.04, 0.1, 0.02),
+  learning_rate = seq(0.1, 0.2, 0.1),
   max_depth = seq(1, 5, 4)
 )
 xgboost_bounds <- list(
   subsample = c(0.2, 1),
   colsample_bytree = c(0.2, 1),
   min_child_weight = c(1L, 10L),
-  learning_rate = c(0.01, 0.3),
+  learning_rate = c(0.1, 0.2),
   max_depth =  c(1L, 10L)
 )
+
+ncores <- ifelse(
+  test = parallel::detectCores() > 4,
+  yes = 4L,
+  no = ifelse(
+    test = parallel::detectCores() < 2L,
+    yes = 1L,
+    no = parallel::detectCores()
+  )
+)
 optim_args <- list(
-  iters.n = 4,
+  iters.n = ncores,
   kappa = 3.5,
   acq = "ucb"
 )
@@ -64,31 +74,31 @@ test_that(
   desc = "test nested cv, bayesian - surv_xgboost_cox",
   code = {
 
-    surv_xgboost_cox_cv <- MLNestedCV$new(
+    surv_xgboost_cox_optimization <- MLNestedCV$new(
       learner = learner,
       strategy = "bayesian",
       fold_list = fold_list,
       k_tuning = 3L,
-      ncores = 2L,
+      ncores = ncores,
       seed = seed
     )
 
-    surv_xgboost_cox_cv$parameter_bounds <- xgboost_bounds
-    surv_xgboost_cox_cv$parameter_grid <- param_list_xgboost
-    surv_xgboost_cox_cv$split_type <- "stratified"
-    surv_xgboost_cox_cv$split_vector <- split_vector
-    surv_xgboost_cox_cv$optim_args <- optim_args
+    surv_xgboost_cox_optimization$parameter_bounds <- xgboost_bounds
+    surv_xgboost_cox_optimization$parameter_grid <- param_list_xgboost
+    surv_xgboost_cox_optimization$split_type <- "stratified"
+    surv_xgboost_cox_optimization$split_vector <- split_vector
+    surv_xgboost_cox_optimization$optim_args <- optim_args
 
     # set data
-    surv_xgboost_cox_cv$set_data(
+    surv_xgboost_cox_optimization$set_data(
       x = train_x,
       y = train_y
     )
 
-    cv_results <- surv_xgboost_cox_cv$execute()
+    cv_results <- surv_xgboost_cox_optimization$execute()
     expect_type(cv_results, "list")
     expect_equal(dim(cv_results), c(5, 9))
-    expect_true(inherits(x = surv_xgboost_cox_cv$results, what = "mlexCV"))
+    expect_true(inherits(x = surv_xgboost_cox_optimization$results, what = "mlexCV"))
   }
 )
 
@@ -97,31 +107,31 @@ test_that(
   desc = "test nested cv, grid - surv_xgboost_cox",
   code = {
 
-    surv_xgboost_cox_cv <- MLNestedCV$new(
+    surv_xgboost_cox_optimization <- MLNestedCV$new(
       learner = learner,
       strategy = "grid",
       fold_list = fold_list,
       k_tuning = 3L,
-      ncores = 2L,
+      ncores = ncores,
       seed = seed
     )
     set.seed(seed)
     random_grid <- sample(1:nrow(param_list_xgboost), 10)
-    surv_xgboost_cox_cv$parameter_grid <- param_list_xgboost[random_grid, ]
-    surv_xgboost_cox_cv$split_type <- "stratified"
-    surv_xgboost_cox_cv$split_vector <- split_vector
-    surv_xgboost_cox_cv$optim_args <- optim_args
+    surv_xgboost_cox_optimization$parameter_grid <- param_list_xgboost[random_grid, ]
+    surv_xgboost_cox_optimization$split_type <- "stratified"
+    surv_xgboost_cox_optimization$split_vector <- split_vector
+    surv_xgboost_cox_optimization$optim_args <- optim_args
 
     # set data
-    surv_xgboost_cox_cv$set_data(
+    surv_xgboost_cox_optimization$set_data(
       x = train_x,
       y = train_y
     )
 
-    cv_results <- surv_xgboost_cox_cv$execute()
+    cv_results <- surv_xgboost_cox_optimization$execute()
     expect_type(cv_results, "list")
     expect_equal(dim(cv_results), c(5, 9))
-    expect_true(inherits(x = surv_xgboost_cox_cv$results, what = "mlexCV"))
+    expect_true(inherits(x = surv_xgboost_cox_optimization$results, what = "mlexCV"))
   }
 )