+++ tests/testthat/test-cv_surv_ranger_cox.R
@@ -2,14 +2,18 @@ dataset <- survival::colon |>
   data.table::as.data.table() |>
   na.omit()
 
-learner <- mlexperiments::LearnerSurvGlmnetCox
+learner <- mlexperiments::LearnerSurvRangerCox
 seed <- 123
 surv_cols <- c("status", "time", "rx")
 
 feature_cols <- colnames(dataset)[3:ncol(dataset)]
 
-param_list_glmnet <- expand.grid(
-  alpha = seq(0, 1, .2)
+param_list_ranger <- expand.grid(
+  sample.fraction = seq(0.6, 1, .2),
+  min.node.size = seq(1, 5, 4),
+  mtry = seq(2, 6, 2),
+  num.trees = c(5L, 10L),
+  max.depth = seq(1, 5, 4)
 )
 
 ncores <- ifelse(
@@ -40,6 +44,7 @@ train_y <- survival::Surv(
   type = "right"
 )
 
+options("mlexperiments.bayesian.max_init" = 10L)
 
 fold_list <- splitTools::create_folds(
   y = split_vector,
@@ -49,26 +54,28 @@ fold_list <- splitTools::create_folds(
 )
 
 test_that(
-  desc = "test cv - surv_glmnet_cox",
+  desc = "test cv - surv_ranger_cox",
   code = {
 
-    surv_glmnet_cox_optimization <- MLCrossValidation$new(
+    surv_ranger_cox_optimization <- MLCrossValidation$new(
       learner = learner,
       fold_list = fold_list,
       ncores = ncores,
       seed = seed
     )
-    surv_glmnet_cox_optimization$learner_args = list(alpha = 0.8, lambda = 0.002)
+    surv_ranger_cox_optimization$learner_args <- as.list(
+      data.table::data.table(param_list_ranger[1, ], stringsAsFactors = FALSE)
+    )
 
     # set data
-    surv_glmnet_cox_optimization$set_data(
+    surv_ranger_cox_optimization$set_data(
       x = train_x,
       y = train_y
     )
 
-    cv_results <- surv_glmnet_cox_optimization$execute()
+    cv_results <- surv_ranger_cox_optimization$execute()
     expect_type(cv_results, "list")
-    expect_equal(dim(cv_results), c(10, 3))
-    expect_true(inherits(x = surv_glmnet_cox_optimization$results, what = "mlexCV"))
+    expect_equal(dim(cv_results), c(10, 6))
+    expect_true(inherits(x = surv_ranger_cox_optimization$results, what = "mlexCV"))
   }
 )